requirements.txt
django-environ==0.10.0
paho-mqtt==1.6.1
prometheus-client==0.17.1
